Put auth login in hook
