+++ extension/app/src/pages/MainPage.tsx
@@ -8,68 +8,11 @@ import {
   Spinner,
   TextArea,
 } from "@dust-tt/sparkle";
-import { useCallback, useEffect, useState } from "react";
 
-import { sendAuthMessage, sentLogoutMessage } from "../lib/auth";
-import {
-  clearAccessToken,
-  getAccessToken,
-  saveAccessToken,
-} from "../lib/utils";
+import { useAuth } from "../hooks/useAuth";
 
 export const MainPage = () => {
-  const [token, setToken] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-
-  useEffect(() => {
-    const fetchToken = async () => {
-      try {
-        const storedToken = await getAccessToken();
-        if (storedToken && typeof storedToken === "string") {
-          setToken(storedToken);
-        }
-      } catch (error) {
-        console.error("Error retrieving token:", error);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-    void fetchToken();
-  }, []);
-
-  const handleLogin = useCallback(async () => {
-    setIsLoading(true);
-    try {
-      const response = await sendAuthMessage();
-      if (response?.accessToken) {
-        await saveAccessToken(response.accessToken);
-        setToken(response.accessToken);
-      } else {
-        console.error("Authentication failed.");
-      }
-    } catch (error) {
-      console.error("Error sending auth message:", error);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [sendAuthMessage, setIsLoading, setToken]);
-
-  const handleLogout = useCallback(async () => {
-    setIsLoading(true);
-    try {
-      const response = await sentLogoutMessage();
-      if (response?.success) {
-        await clearAccessToken();
-        setToken(null);
-      } else {
-        console.error("Logout failed.");
-      }
-    } catch (error) {
-      console.error("Error sending logout message:", error);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [setToken]);
+  const { token, isLoading, handleLogin, handleLogout } = useAuth();
 
   return (
     <div className="flex flex-col p-4 gap-2 h-screen">
